Types fixes
