+++ aiopg/py.typed
@@ -1,0 +1,1 @@
+Marker
