+++ aiopg/sa/connection.py
@@ -6,7 +6,7 @@ from sqlalchemy.sql import ClauseElement
 from sqlalchemy.sql.ddl import DDLElement
 from sqlalchemy.sql.dml import UpdateBase
 
-from ..utils import _SAConnectionContextManager, _TransactionContextManager
+from ..utils import _ContextManager
 from . import exc
 from .result import ResultProxy
 from .transaction import (
@@ -79,7 +79,7 @@ class SAConnection:
 
         """
         coro = self._execute(query, *multiparams, **params)
-        return _SAConnectionContextManager(coro)
+        return _ContextManager[SAConnection](coro, lambda x: x.close())
 
     async def _open_cursor(self):
         if self._connection is None:
@@ -204,15 +204,16 @@ class SAConnection:
 
         """
         coro = self._begin(isolation_level, readonly, deferrable)
-        return _TransactionContextManager(coro)
+        return _ContextManager[Transaction](
+            coro, self._commit_if_active, self._rollback
+        )
 
     async def _begin(self, isolation_level, readonly, deferrable):
         if self._transaction is None:
             self._transaction = RootTransaction(self)
             await self._begin_impl(isolation_level, readonly, deferrable)
             return self._transaction
-        else:
-            return Transaction(self, self._transaction)
+        return Transaction(self, self._transaction)
 
     async def _begin_impl(self, isolation_level, readonly, deferrable):
         stmt = 'BEGIN'
@@ -261,7 +262,7 @@ class SAConnection:
         transaction of a whole.
         """
         coro = self._begin_nested()
-        return _TransactionContextManager(coro)
+        return _ContextManager(coro, self._commit_if_active, self._rollback)
 
     async def _begin_nested(self):
         if self._transaction is None:
@@ -389,6 +390,15 @@ class SAConnection:
                 await self._engine.release(self)
             self._connection = None
             self._engine = None
+
+    @staticmethod
+    async def _commit_if_active(t: Transaction) -> None:
+        if t.is_active:
+            await t.commit()
+
+    @staticmethod
+    async def _rollback(t: Transaction) -> None:
+        await t.rollback()
 
 
 def _distill_params(multiparams, params):
